+++ changelog.md
@@ -3,6 +3,7 @@
 ## 1.4.2 (TBD)
 
 - Fixed IndexOutOfRange exception caused by trying to add / remove goop circles outside the bounds of the current map
+- Added "Optimize Pause Menu" option, dramatically reducing CPU usage while the game is paused
 
 ## 1.4.1 (2025-06-07)